♻️[Refactor] #16 - 코드 구조 리팩토링
